Add XVTR support

+++ SDR_RA8875/Display.h
@@ -11,7 +11,7 @@
 uint16_t grandient(uint8_t val);
 void draw_2_state_Button(uint8_t button, uint8_t *function_ptr);
 void refreshScreen(void);
-const char * formatVFO(uint32_t vfo);
+const char * formatVFO(uint64_t vfo);
 void displayTime(void);
 void displayMeter(int val, const char *string, uint16_t colorscheme);
 void drawLabel(uint8_t lbl_num, uint8_t *function_ptr);
